Refactor
Still the same functionality as the previous version, but code quality
has been greatly improved. The handle_job function is now broken into
several pieces that can be (should be, have not done so yet) tested
independently.

Nearing merge-ready, just having a look tomorrow as well to see if
there's anything more I need to fix. Also need to try this in concert
with the rest of the system.