Yet another big refactor
Further refactor, and fixed a lot of small issues with the last commit.
System is now fully tested as working in context. Ready for prod!

The job handling logic in the crud module is greatly simplified, and now
follows a fairly standard workflow: Processing the list of jobs is
basically just a list processing task, now there's no more raising and
recursing, just list-munching.

Also fixed a small error relating to error propagation: Forgot to decode
the HTTP response content with the error message.