Retry on 503
Timeout errors happen downstream. This is unfortunate, but a fact of
life when you're working with an underpowered server.

Before, this used to blow up in the users' face, but it's much nicer to
keep these kinds of issues away from users, instead retrying. This works
well, as the server is able to handle most tasks on second-try, after
getting other tasks out of the way.

To support this, I wrote an Error model that can be shared around as
JSON in redis. This model lets the app case on HTTP code, retrying 503s
instead of "raising" them back to the user.

Works really well, this change makes the app able to handle a lot more
traffic than before!